+++ probe-rs/src/memory/mod.rs
@@ -1,4 +1,5 @@
 use crate::error::Error;
+use crate::SubArray;
 
 use anyhow::{anyhow, Result};
 use scroll::Pread;
@@ -221,7 +222,7 @@ pub trait MemoryInterface {
 
         let mut buffer = vec![0u32; inbetween_count / 4];
         for (bytes, value) in data.chunks_exact(4).zip(buffer.iter_mut()) {
-            *value = u32::from_le_bytes([bytes[0], bytes[1], bytes[2], bytes[3]]);
+            *value = u32::from_le_bytes(*bytes.subarray());
         }
         self.write_32(address, &buffer)?;